+++ aim/artifacts/metric.py
@@ -4,8 +4,9 @@ from typing import Union, Optional
 
 from aim.artifacts import Record
 from aim.artifacts.artifact import Artifact
+from aim.artifacts.proto.base_pb2 import BaseRecord
 from aim.artifacts.proto.metric_pb2 import MetricRecord
-from aim.artifacts.utils import validate_dict
+from aim.artifacts.utils import validate_mapping, validate_iterable
 
 
 class Metric(Artifact):
@@ -15,38 +16,52 @@ class Metric(Artifact):
                  value: Union[int, float],
                  epoch: int = None,
                  step: int = None,
-                 aim_session_id: Optional[int] = None,
                  **kwargs):
         if not self.validate_name(str(name)):
-            raise ValueError('metric name can contain only letters, numbers, ' +
-                             'underscore, dash and space')
+            raise ValueError('metric name must be a python identifier')
+        self.name = str(name)
 
         if not isinstance(value, (int, float)):
             raise TypeError('metric value must be a type of int or float')
+        self.value = value
 
-        val_res, val_item = validate_dict(
-            kwargs, (str, int, float,),
-            (str, int, float, bool,))
-        if not val_res:
-            raise TypeError(('Metric context contains illegal item: '
+        # Construct context kwargs
+        context_kwargs = {}
+        for k, v in kwargs.items():
+            if not k.startswith('__'):
+                context_kwargs[k] = v
+
+        # Validate context
+        val_res, val_item = validate_mapping(
+            context_kwargs,
+            (str,),
+            (str, int, float, tuple, bool, type(None)),
+            key_str_validator=r'^[^\d\W]\w*\Z',
+            iterable_validator=lambda x: validate_iterable(x,
+                                                           (str, int, float,
+                                                            tuple, bool,
+                                                            type(None)))
+        )
+        if val_res == 1:
+            raise TypeError(('metric context contains illegal item: ' +
                              '`{}` of type `{}`').format(val_item,
                                                          type(val_item)))
-
-        self.name = re.sub(' +', ' ', str(name))
-        self.value = value
-        self.epoch = epoch
-
-        self.context = kwargs if len(kwargs.keys()) else {}
-        self.hashable_context = tuple(sorted(kwargs.items()))
+        elif val_res == 2:
+            raise TypeError(('metric context key must be a ' +
+                             'python identifier: `{}`').format(val_item))
+        self.context = context_kwargs if len(context_kwargs.keys()) else {}
+        self.hashable_context = tuple(sorted(context_kwargs.items()))
 
         super(Metric, self).__init__(self.cat)
+
+        self.epoch = epoch
 
         if self.context is not None:
             step_meta = self.hashable_context
         else:
             step_meta = None
         self.initialize_step_counter(step, self.name, step_meta,
-                                     session_id=aim_session_id)
+                                     session_id=kwargs.get('__aim_session_id'))
 
     def __str__(self):
         return '{name}: {value}'.format(name=self.name,
@@ -65,6 +80,16 @@ class Metric(Artifact):
             context=self.context,
             binary_type=self.PROTOBUF,
         )
+
+    @staticmethod
+    def deserialize_pb(data):
+        base_pb = BaseRecord()
+        base_pb.ParseFromString(data)
+
+        metric_pb = MetricRecord()
+        metric_pb.ParseFromString(base_pb.artifact)
+
+        return base_pb, metric_pb
 
     def save_blobs(self, name: str, abs_path: str = None):
         pass