+++ aim/cli/de/utils.py
@@ -1,9 +1,11 @@
 import click
 
+from aim.engine.container import AimContainer
+
 
 def repo_init_alert():
     click.echo('Repository not found')
-    click.echo('    (use "aim init" to initialize empty repository)')
+    click.echo('    (use "aim init" to initialize a new repository)')
 
 
 def docker_requirement_alert():
@@ -17,3 +19,10 @@ def docker_image_pull_fail_alert():
     click.echo(
         click.style('    (NOTE: If somehow docker is stuck downloading, ' +
                     'please restart docker and try again)', fg='yellow'))
+
+
+def check_docker_dependency(exit_process=True):
+    if not AimContainer.is_docker_installed():
+        docker_requirement_alert()
+        if exit_process:
+            exit()